refactor: VoteReportRepository 인터페이스 적용
