+++ src/main/java/com/mjuAppSW/joA/domain/voteReport/repository/VoteReportRepository.java
@@ -1,0 +1,11 @@
+package com.mjuAppSW.joA.domain.voteReport.repository;
+
+import com.mjuAppSW.joA.domain.voteReport.VoteReport;
+import java.util.Optional;
+
+public interface VoteReportRepository {
+
+    void save(VoteReport voteReport);
+
+    Optional<VoteReport> findBy(Long voteId);
+}
